+++ youtube_scraper/index.js
@@ -1,0 +1,25 @@
+const puppeteer = require('puppeteer');
+const youtube = require('./youtube');
+
+try {
+    (async () => {
+
+        let keywords = [
+            'history of footwear',
+            'the evolution of footwear'
+        ];
+
+        const browser = await puppeteer.launch();
+        let results = await youtube.scrape_youtube(browser, keywords);
+        // let link = await youtube.scrape_youtube(browser,keywords)
+        console.dir(results, { depth: null, colors: true });
+        let history =`https://youtube.com${results['history of footwear'].results[0].link}`;
+        console.log(history)
+        let evolution = `https://youtube.com${results['the evolution of footwear'].results[0].link}`
+        console.log(evolution)
+        await browser.close();
+
+    })()
+} catch (err) {
+    console.error(err)
+}
